Feedback points on login otp
When the input the box is focussed then the error message should be disappeared: Not a blocker to go live
Submit button will be changed to Back to Login and on click of it it will redirect to the login page when customer email is already registered with another number: Blocker to go live.
Add Padding in the error message: Not a blocker to go live.
Clear the input field on click when there is any error: Not a blocker to go live.